StateEligibility.js: Refactor to editable strings
Use an Array of Strings (actually an Array of Objects of Strings) to
generate the <ListItem/>s, instead of manually entering all of them.
Now editing the criteria is much simpler and more centralized.

Remove the EligibilityGroupItem abstraction for ListItem, it was
unnecessary and seperates the icon from the text, making it harder to
parametrize the icon in the future.

Handle the goofiness that sometimes we have a title <div> and
sometimes we do not.

Unfortunately we have to add keys to dynamically generated react
fragments to quell warnings about duplicate keys, which means we have to use the
<React.Fragment key="xyz"> ... </React.Fragment>
syntax rather than the more common abbreviated
<> ... </>
syntax. Hopefully this isn't too confusing.