StateEligibility.js: Use PeopleIcon instead of CheckIcon
As noted by me and confirmed by Hal, the CheckIcon had an affordance
for completion, the implication that you're supposed to do all the
checkable items. Although it's not seriously contemplated that anyone
would actually *think* that, it's the wrong application for a checkbox
icon.

Although we originally discussed using a bullet icon, material UI
doesn't seem to have a good choice -- there is
RadioButtonUncheckedIcon but it's a fairly large circle.

Instead, let's use icons that actually represent groups of people,
which is what we are listing anyhow. That means...PeopleIcon!